PapyrusDotNet v1.0 TP2f1
Added a new input argument ( -noui ) for PapyrusDotNet that can be used to prevent the application from crashing when using it in a post build event.
Moved the Inversion of control container (IoCContainer) to the shared PapyrusDotNet.Common project instead. This is now being used in PapyrusDotNet to resolve some of the classes used. Idea is to get everything resolved using the IoCContainer so implementation can be switched with breeze.